+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Updated
 - Upgraded to Gravity `v0.10.0` and updated header & footer HTML accordingly
 
+### Fixed
+- WEB-261: Math.hypot fallback for IE in hero
+
 
 ## [4.0.2] - 2018-05-18
 ### Added